added Root Endpoint
added the root endpoint to return some stuff, like the version of the KitsuDeck firmware (have to find out how to use .env), if its protected by a password(needed to add it to the dash, for future ws requests).

I will have to figure out how i will store the makros on the keyboard. I will probably look if i can use sqlite or some sort of database.