added storage driver + png/jpg/bmp decode
right now the KitsuDeck will load all registered images on the `DeckCol`

In the next step i will finish the macro register function in the KitsuDeckDash program, so the deck will load macro btns and not just images. I think there will be a limit of 25 macros per site, due to the time the esp32 needs to load/decode the png images. I had it on 48 per site before, but than it took about 10 sek to load the home site where the macros are.